Move compressed word into primitives
