+++ library/Rattletrap/Primitive/CompressedWord.hs
@@ -1,4 +1,4 @@
-module Rattletrap.CompressedWord where
+module Rattletrap.Primitive.CompressedWord where
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
